+++ src/test/php/web/frontend/unittest/HandlebarsTest.class.php
@@ -121,4 +121,28 @@ class HandlebarsTest {
   public function current_date($template) {
     Assert::equals(Date::now()->toString('d.m.Y'), $this->transform($template));
   }
+
+  #[Test, Values([['', 0], ['null', 0], ['"test"', 1], ['null "test"', 1], ['numbers', 1], ['empty', 0]])]
+  public function any($expr, $expected) {
+    Assert::equals((string)$expected, $this->transform('{{any '.$expr.'}}', [
+      'numbers' => [1, 2, 3],
+      'empty'   => [],
+    ]));
+  }
+
+  #[Test, Values([['', 1], ['null', 1], ['"test"', 0], ['null "test"', 0], ['numbers', 0], ['empty', 1]])]
+  public function none($expr, $expected) {
+    Assert::equals((string)$expected, $this->transform('{{none '.$expr.'}}', [
+      'numbers' => [1, 2, 3],
+      'empty'   => [],
+    ]));
+  }
+
+  #[Test, Values([['', 0], ['null', 0], ['"test"', 1], ['null "test"', 0], ['numbers', 1], ['empty', 0]])]
+  public function all($expr, $expected) {
+    Assert::equals((string)$expected, $this->transform('{{all '.$expr.'}}', [
+      'numbers' => [1, 2, 3],
+      'empty'   => [],
+    ]));
+  }
 }