+++ ChangeLog.md
@@ -3,6 +3,7 @@ Handlebars for XP web frontends change log
 
 ## ?.?.? / ????-??-??
 
+* Add `min` and `max` helpers - @thekid
 * Add `contains` helper - @thekid
 
 ## 0.1.0 / 2021-02-13