+++ ChangeLog.md
@@ -3,6 +3,8 @@ Handlebars for XP web frontends change log
 
 ## ?.?.? / ????-??-??
 
+## 0.2.0 / 2021-02-13
+
 * Add `any`, `none` and `all` helpers - @thekid
 * Add `min` and `max` helpers - @thekid
 * Add `contains` helper - @thekid